Unix time in uint32_t, not uint64_t
uint32_t suffices for ~100 years, which should suffice for a personal
project...

+++ light.h
@@ -145,7 +145,7 @@ class WakeupLight {
 
   private:
     bool inWakeup() {
-      uint64_t now = Clock.timestamp();
+      uint32_t now = Clock.timestamp();
       if (now == 0) {
         // Clock is not yet initialized.
         return false;
